+++ database/migrations/2015_03_11_143941_create_syllabary_row_header_table.php
@@ -14,11 +14,11 @@ class CreateSyllabaryRowHeaderTable extends Migration {
 	{
 		Schema::create('syllabary_row_header', function(Blueprint $table)
 		{
-			$table->increments('id');
-			$table->integer('syllabary_id')->unsigned();
+			$table->bigIncrements('id');
+			$table->bigInteger('syllabary_id')->unsigned();
 			$table->string('ipa', 10);
-      $table->integer('symbol_id')->unsigned();
-			$table->integer('index')->unsigned();
+      		$table->bigInteger('symbol_id')->unsigned();
+			$table->bigInteger('index')->unsigned();
 			$table->timestamps();
 		});
 	}
